Move findbatch test to MongoClientTest because MongoClientTestBase is called from VerticleTest and findBatch doesnt work with ProxyHandler
