+++ vertx-mongo-client/src/test/java/io/vertx/ext/mongo/MongoClientTest.java
@@ -1,6 +1,7 @@
 package io.vertx.ext.mongo;
 
 import io.vertx.core.json.JsonObject;
+import org.junit.Test;
 
 import java.util.concurrent.CountDownLatch;
 
@@ -25,4 +26,21 @@ public class MongoClientTest extends MongoClientTestBase {
     super.tearDown();
   }
 
+
+  @Test
+  public void testFindBatch() throws Exception {
+    int numDocs = 200;
+
+    String collection = randomCollection();
+    CountDownLatch latch = new CountDownLatch(numDocs);
+    mongoClient.createCollection(collection, onSuccess(res -> {
+      insertDocs(collection, numDocs, onSuccess(res2 -> {
+        mongoClient.findBatchWithOptions(collection, new JsonObject(), new FindOptions(), onSuccess(result -> {
+          assertNotNull(result);
+          latch.countDown();
+        }));
+      }));
+    }));
+    awaitLatch(latch);
+  }
 }
